Separate AWS
Separate AWS actions from main file. Changes string injection in certain places to `os.path.join` when it makes sense.
